refactor page folder
