+++ src/pages/home-page/artist-section/ArtistSection.css
