+++ src/pages/home-page/artist-section/ArtistSection.tsx
@@ -2,8 +2,8 @@ import CreateNewFolderOutlinedIcon from "@mui/icons-material/CreateNewFolderOutl
 import FavoriteBorderIcon from "@mui/icons-material/FavoriteBorder";
 import ShareOutlinedIcon from "@mui/icons-material/ShareOutlined";
 import { Box, IconButton } from "@mui/material";
-import TrackTable from "../../components/TrackTable";
-import "../styles/ArtistPage.css";
+import TrackTable from "../../../components/TrackTable";
+import "./ArtistSection.css";
 
 const url = "/short-adventurous-intro-1-117090.mp3";
 
@@ -30,7 +30,7 @@ const sampleTracks = [
   },
 ];
 
-export default function ArtistPage() {
+export default function ArtistSection() {
   return (
     <Box className="artist-content">
       <Box className="artist-top-section">
