+++ src/pages/home-page/download-history-section/DownloadHistory.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 const DownloadHistory: React.FC = () => {
-  return <div>Welcome to the DownloadHistory page!</div>;
+  return <div>Welcome to the DownloadHistory Section!</div>;
 };
 
 export default DownloadHistory;
