+++ src/pages/home-page/favorite-section/Favorite.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 const Favorite: React.FC = () => {
-  return <div>Welcome to the Favorite page!</div>;
+  return <div>Welcome to the Favorite Section!</div>;
 };
 
 export default Favorite;
